+++ src/main/java/dinner/dev/endless_gravity/EndlessGravity.java
@@ -1,12 +1,9 @@
 package dinner.dev.endless_gravity;
 
 import dinner.dev.endless_gravity.block.ModBlocks;
-import dinner.dev.endless_gravity.block.StarshipBlock;
 import dinner.dev.endless_gravity.compat.sable.AtmosphereForceGroups;
 import dinner.dev.endless_gravity.item.ModItems;
 import dinner.dev.endless_gravity.network.ConfigSyncPayload;
-import dinner.dev.endless_gravity.network.RocketSoundPayload;
-import dinner.dev.endless_gravity.particle.ModParticles;
 import dinner.dev.endless_gravity.sound.ModSounds;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
@@ -18,7 +15,6 @@ import net.neoforged.neoforge.network.PacketDistributor;
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
 import net.neoforged.neoforge.network.registration.PayloadRegistrar;
 import net.minecraft.server.level.ServerPlayer;
-import net.neoforged.neoforge.event.tick.ServerTickEvent;
 import org.slf4j.Logger;
 import com.mojang.logging.LogUtils;
 
@@ -33,18 +29,12 @@ public class EndlessGravity {
         AtmosphereForceGroups.init(modEventBus);
         ModItems.init(modEventBus);
         ModSounds.init(modEventBus);
-        ModParticles.init(modEventBus);
         modEventBus.addListener(RegisterPayloadHandlersEvent.class, event -> {
             PayloadRegistrar registrar = event.registrar("1");
             registrar.playToClient(
                     ConfigSyncPayload.ID,
                     ConfigSyncPayload.STREAM_CODEC,
                     ConfigSyncPayload::handle
-            );
-            registrar.playToClient(
-                    RocketSoundPayload.ID,
-                    RocketSoundPayload.STREAM_CODEC,
-                    RocketSoundPayload::handle
             );
         });
 
@@ -53,9 +43,6 @@ public class EndlessGravity {
                 PacketDistributor.sendToPlayer(serverPlayer, ConfigSyncPayload.fromConfig());
             }
         });
-
-        NeoForge.EVENT_BUS.addListener(ServerTickEvent.Pre.class,
-                event -> StarshipBlock.onServerTick(event.getServer()));
 
         LOGGER.info("Endless Gravity loaded");
     }